backend/config.py
import os
from dotenv import load_dotenv

"""
Create a .env file to store these values
"""
load_dotenv()

# Ollama configuration
OLLAMA_URL = os.getenv("OLLAMA_URL")
OLLAMA_MODEL = os.getenv("OLLAMA_MODEL")

# Groq configuration
GROQ_API_KEY = os.getenv("GROQ_API_KEY")
GROQ_MODEL = os.getenv("GROQ_MODEL")
