backend/requirements.txt
fastapi[all]
uvicorn
youtube-transcript-api
requests
pydantic
python-dotenv
groq
ollama
tiktoken